LayoutTests/accessibility/ios-simulator/disabled-states-expected.txt

button 1
button 2
button 3

This test thats disabled and aria-disabled change the traits returned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS btn1.traits != btn2.traits is true
PASS btn2.traits == btn3.traits is true
PASS select1.traits != select2.traits is true
PASS successfullyParsed is true

TEST COMPLETE
